validate travel pass card name against config
