+++ api/config/travel-pass.yml
@@ -1,0 +1,3 @@
+authorized:
+  - navigo
+
