+++ api/shared/providers/mongo/schemas/getters.js
@@ -1,0 +1,4 @@
+/**
+ * Mongoose getters
+ */
+module.exports = {};
