+++ api/shared/providers/mongo/schemas/regex.js
@@ -1,0 +1,14 @@
+module.exports = {
+  phone: /^(?:(?:\+)[0-9]{1,3}|0)(\s*\(0\))?\s*[1-9](?:[\s.-]*\d{2}){4}$/,
+  email: /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/,
+  insee: /^[0-9][0-9AB][0-9]{3}$/,
+  siren: /^[0-9]{9}$/,
+  siret: /^[0-9]{12}$/,
+  naf: /^[0-9]{4}[A-Z]{1}$/,
+  nic: /^[0-9]{5}$/,
+  vatIntra: /^[A-Z]{2}[A-Z0-9]{2}[0-9]{9}$/,
+  postcode: /^[0-9][0-9AB][0-9]{2,3}$/,
+  cedex: /^[0-9]{5}$/,
+  lon: /^-?[0-9]{1,3}(\.[0-9]+)?$/,
+  lat: /^-?[0-9]{1,2}(\.[0-9]+)?$/,
+};
